Rename base factory to createBasePropsBuilder
Free up the name createPropsBuilder so a future platform-resolved
factory can claim it. The base factory is now internal-only (no longer
re-exported from common/style/index.ts).

- Rename file: common/style/createPropsBuilder.ts -> createBasePropsBuilder.ts
- Rename default export: createPropsBuilder -> createBasePropsBuilder
- Update internal imports in common/style/propsBuilder.ts and
  common/web/style/{propsBuilder,ruleBuilder}.ts to import the base
  directly via deep import.
- Drop the createPropsBuilder re-export from common/style/index.ts.
- Rename the test file accordingly.

No behavior change.
