+++ packages/react-native-reanimated/src/common/style/index.ts
@@ -1,6 +1,6 @@
 'use strict';
 export * from './config';
-export { default as createPropsBuilder } from './createPropsBuilder';
+export { default as createBasePropsBuilder } from './createBasePropsBuilder';
 export * from './processors';
 export { type NativePropsBuilder, stylePropsBuilder } from './propsBuilder';
 export * from './registry';
